+++ src/app/(main)/(library)/_containers/local-library/local-library.tsx
@@ -5,12 +5,13 @@ import {
     completed_libraryEntryAtoms,
     currentlyWatching_libraryEntryAtoms,
     LibraryEntry,
+    paused_libraryEntryAtoms,
+    planning_libraryEntryAtoms,
     rest_libraryEntryAtoms,
 } from "@/atoms/library/library-entry.atoms"
 import { Atom } from "jotai"
 import { useSelectAtom } from "@/atoms/helpers"
 import { AnimeListItem } from "@/components/shared/anime-list-item"
-import { Divider } from "@/components/ui/divider"
 import { useAtomValue } from "jotai/react"
 import { Slider } from "@/components/shared/slider"
 import { ContinueWatching } from "@/app/(main)/(library)/_containers/continue-watching/continue-watching"
@@ -18,11 +19,13 @@ import { ContinueWatching } from "@/app/(main)/(library)/_containers/continue-wa
 export function LocalLibrary() {
 
     const currentlyWatchingEntryAtoms = useAtomValue(currentlyWatching_libraryEntryAtoms)
+    const pausedEntryAtoms = useAtomValue(paused_libraryEntryAtoms)
+    const planningEntryAtoms = useAtomValue(planning_libraryEntryAtoms)
     const restEntryAtoms = useAtomValue(rest_libraryEntryAtoms)
     const completedEntryAtoms = useAtomValue(completed_libraryEntryAtoms)
 
     return (
-        <div className={"p-4 space-y-8"}>
+        <div className={"p-4 space-y-8 relative z-[5]"}>
             {currentlyWatchingEntryAtoms.length > 0 && <>
                 <h2>Continue watching</h2>
                 <Slider>
@@ -33,7 +36,7 @@ export function LocalLibrary() {
                         />
                     })}
                 </Slider>
-                <Divider/>
+                {/*<Divider/>*/}
                 <h2>Currently watching</h2>
                 <div
                     className={"grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5 2xl:grid-cols-7 min-[2000px]:grid-cols-8 gap-4"}>
@@ -41,16 +44,41 @@ export function LocalLibrary() {
                         return <EntryAnimeItem key={`${entryAtom}`} entryAtom={entryAtom}/>
                     })}
                 </div>
-                <Divider/>
+                {/*<Divider/>*/}
             </>}
-            <div
-                className={"grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5 2xl:grid-cols-7 min-[2000px]:grid-cols-8 gap-4"}>
-                {restEntryAtoms.map(entryAtom => {
-                    return <EntryAnimeItem key={`${entryAtom}`} entryAtom={entryAtom}/>
-                })}
-            </div>
+            {pausedEntryAtoms.length > 0 && <>
+                <h2>Paused</h2>
+                <div
+                    className={"grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5 2xl:grid-cols-7 min-[2000px]:grid-cols-8 gap-4"}
+                >
+                    {pausedEntryAtoms.map(entryAtom => {
+                        return <EntryAnimeItem key={`${entryAtom}`} entryAtom={entryAtom}/>
+                    })}
+                </div>
+                {/*<Divider/>*/}
+            </>}
+            {planningEntryAtoms.length > 0 && <>
+                <h2>Planned</h2>
+                <div
+                    className={"grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5 2xl:grid-cols-7 min-[2000px]:grid-cols-8 gap-4"}
+                >
+                    {planningEntryAtoms.map(entryAtom => {
+                        return <EntryAnimeItem key={`${entryAtom}`} entryAtom={entryAtom}/>
+                    })}
+                </div>
+                {/*<Divider/>*/}
+            </>}
+            {restEntryAtoms.length > 0 && <>
+                <div
+                    className={"grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5 2xl:grid-cols-7 min-[2000px]:grid-cols-8 gap-4"}
+                >
+                    {restEntryAtoms.map(entryAtom => {
+                        return <EntryAnimeItem key={`${entryAtom}`} entryAtom={entryAtom}/>
+                    })}
+                </div>
+                {/*<Divider/>*/}
+            </>}
             {completedEntryAtoms.length > 0 && <>
-                <Divider/>
                 <h2>Completed</h2>
                 <div
                     className={"grid grid-cols-2 md:grid-cols-3 lg:grid-cols-4 xl:grid-cols-5 2xl:grid-cols-7 min-[2000px]:grid-cols-8 gap-4"}>