+++ src/components/shared/large-episode-list-item.tsx
@@ -6,31 +6,40 @@ import { imageShimmer } from "@/components/shared/image-helpers"
 
 type LargeEpisodeListItemProps = {
     title: React.ReactNode
-    actionIcon?: React.ReactElement
+    actionIcon?: React.ReactElement | null
     image?: string | null
     onClick?: () => void
     topTitle?: string | null
     meta?: string | null
+    larger?: boolean
 }
 
 export const LargeEpisodeListItem: React.FC<LargeEpisodeListItemProps & Omit<React.ComponentPropsWithoutRef<"div">, "title">> = (props) => {
 
     const {
         children,
-        actionIcon = <AiFillPlayCircle/>,
+        actionIcon = props.actionIcon !== null ? <AiFillPlayCircle/> : undefined,
         image,
         onClick,
         topTitle,
         meta,
         title,
+        larger = false,
         ...rest
     } = props
 
     return <>
         <div
-            className={"rounded-md border border-[--border] overflow-hidden aspect-[4/2] w-96 relative flex items-end flex-none group/missed-episode-item"}
+            className={cn(
+                "rounded-md border border-[--border] overflow-hidden aspect-[4/2] w-96 relative flex items-end flex-none group/missed-episode-item cursor-pointer",
+                {
+                    "w-[30rem]": larger,
+                },
+            )}
+            onClick={onClick}
+            {...rest}
         >
-            <div className={"absolute w-full h-full rounded-md rounded-b-none overflow-hidden z-[1]"}>
+            <div className={"absolute w-full h-full overflow-hidden z-[1]"}>
                 {!!image ? <Image
                     src={image}
                     alt={""}
@@ -46,10 +55,10 @@ export const LargeEpisodeListItem: React.FC<LargeEpisodeListItemProps & Omit<Rea
                 />
             </div>
             <div className={cn(
-                "group-hover/missed-episode-item:opacity-100 text-7xl text-gray-200",
-                "cursor-pointer opacity-0 transition-opacity bg-gray-900 bg-opacity-50 backdrop-blur-md z-[6] absolute w-full h-full flex items-center justify-center",
-            )} onClick={onClick}>
-                {actionIcon}
+                "group-hover/missed-episode-item:opacity-100 text-6xl text-gray-200",
+                "cursor-pointer opacity-0 transition-opacity bg-gray-900 bg-opacity-50 backdrop-blur-md z-[2] absolute w-full h-full flex items-center justify-center",
+            )}>
+                {actionIcon && actionIcon}
             </div>
             <div className={"relative z-[3] w-full p-4 space-y-1"}>
                 {topTitle && <p className={"w-[80%] line-clamp-1 text-[--muted] font-semibold"}>{topTitle}</p>}