+++ src/lib/anify/helpers.ts
@@ -1,18 +1,17 @@
-import { useQuery } from "@tanstack/react-query"
-import { getAnifyAnimeMetadata } from "@/lib/anify/actions"
-
+// FIXME
 export function useAnifyAnimeMetadata(mediaId: number) {
-    const { data } = useQuery({
-        queryKey: ["anify-episode-covers", mediaId],
-        queryFn: async () => {
-            const res = await getAnifyAnimeMetadata(mediaId)
-            if (!!res && Array.isArray(res))
-                return res
-            return []
-        },
-        keepPreviousData: true,
-        cacheTime: 1000 * 60 * 10,
-    })
+    // const { data } = useQuery({
+    //     queryKey: ["anify-episode-covers", mediaId],
+    //     queryFn: async () => {
+    //         const res = await getAnifyAnimeMetadata(mediaId)
+    //         if (!!res && Array.isArray(res))
+    //             return res
+    //         console.log(res)
+    //         return []
+    //     },
+    //     keepPreviousData: false,
+    //     cacheTime: 0,
+    // })
 
-    return { anifyEpisodeData: data }
+    return { anifyEpisodeData: [] }
 }